+++ src/main/java/owl/automaton/algorithms/LanguageMembership.java
@@ -1,0 +1,112 @@
+/*
+ * Copyright (C) 2016 - 2018  (See AUTHORS)
+ *
+ * This file is part of Owl.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package owl.automaton.algorithms;
+
+import java.util.BitSet;
+import java.util.Set;
+import java.util.stream.Collectors;
+import org.immutables.value.Value;
+import owl.automaton.AbstractCachedStatesAutomaton;
+import owl.automaton.Automaton;
+import owl.automaton.EdgesAutomatonMixin;
+import owl.automaton.UltimatelyPeriodicWord;
+import owl.automaton.acceptance.OmegaAcceptance;
+import owl.automaton.edge.Edge;
+import owl.automaton.util.AnnotatedState;
+import owl.factories.ValuationSetFactory;
+import owl.util.annotation.HashedTuple;
+
+public final class LanguageMembership {
+
+  private LanguageMembership() {
+  }
+
+  public static <S, A extends OmegaAcceptance> boolean contains(Automaton<S, A> automaton,
+    UltimatelyPeriodicWord word) {
+    return !LanguageEmptiness.isEmpty(new IndexedAutomaton<>(automaton, word));
+  }
+
+  @Value.Immutable
+  @HashedTuple
+  abstract static class IndexedState<S> implements AnnotatedState<S> {
+    abstract int index();
+
+    static <S> IndexedState<S> of(int index, S state) {
+      return IndexedStateTuple.create(index, state);
+    }
+  }
+
+  private static class IndexedAutomaton<S, A extends OmegaAcceptance>
+    extends AbstractCachedStatesAutomaton<IndexedState<S>, A>
+    implements EdgesAutomatonMixin<IndexedState<S>, A> {
+    private final Automaton<S, A> automaton;
+    private final UltimatelyPeriodicWord word;
+
+    private IndexedAutomaton(Automaton<S, A> automaton,
+      UltimatelyPeriodicWord word) {
+      this.automaton = automaton;
+      this.word = word;
+    }
+
+    @Override
+    public A acceptance() {
+      return automaton.acceptance();
+    }
+
+    @Override
+    public ValuationSetFactory factory() {
+      return automaton.factory();
+    }
+
+    @Override
+    public Set<IndexedState<S>> initialStates() {
+      return automaton.initialStates().stream()
+        .map(x -> IndexedStateTuple.create(-word.prefix.size(), x))
+        .collect(Collectors.toUnmodifiableSet());
+    }
+
+    @Override
+    public Set<Edge<IndexedState<S>>> edges(IndexedState<S> state, BitSet valuation) {
+      BitSet allowedValuation;
+
+      if (state.index() < 0) {
+        allowedValuation = word.prefix.get(-state.index() - 1);
+      } else {
+        allowedValuation = word.period.get(state.index());
+      }
+
+      if (!allowedValuation.equals(valuation)) {
+        return Set.of();
+      }
+
+      Set<Edge<S>> edges = automaton.edges(state.state(), valuation);
+      int nextIndex;
+      if (state.index() + 1 >= word.period.size()) {
+        nextIndex = (state.index() + 1) % word.period.size();
+      } else {
+        nextIndex = state.index() + 1;
+      }
+
+      return edges.stream()
+        .map(x -> x.withSuccessor(IndexedState.of(nextIndex, x.successor())))
+        .collect(Collectors.toUnmodifiableSet());
+    }
+  }
+}